requirements.txt
requests
boto3
psycopg2-binary
pyspark
python-dotenv